Process more reliably files filled with zeros
Until now, we would entirely rely on xxhash().
However when some files of different sizes are not empty yet entirely
filled with zeros (e.g. corrupted files), the xxhash() would be
different depending on the file size. The hashing function now returns 0
(regardless of the file size) when the hashed data is full of zeros.

A bug is also fixed in dupfiles (that would otherwise only check the
hash without looking whether results have identical size)